pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nbdet"
version = "0.1.0"
description = "News summarizer and political bias detection"
dependencies = [
    "requests",
    "python-dotenv",
    "transformers",
    "torch",
    "streamlit"
]
